Make document export a service
