+++ mpvqc/injections.py
@@ -19,13 +19,14 @@ import inject
 
 from mpvqc.services import ResourceService, ApplicationEnvironmentService, ApplicationPathsService, BackupService, \
     FileStartupService, PlayerService, ResourceReaderService, ReverseTranslatorService, \
-    OperatingSystemZoomDetectorService, FontLoaderService, SettingsService
+    OperatingSystemZoomDetectorService, FontLoaderService, SettingsService, DocumentExporterService
 
 
 def bindings(binder: inject.Binder):
     binder.bind_to_constructor(ApplicationEnvironmentService, lambda: ApplicationEnvironmentService())
     binder.bind_to_constructor(ApplicationPathsService, lambda: ApplicationPathsService())
     binder.bind_to_constructor(BackupService, lambda: BackupService())
+    binder.bind_to_constructor(DocumentExporterService, lambda: DocumentExporterService())
     binder.bind_to_constructor(FileStartupService, lambda: FileStartupService())
     binder.bind_to_constructor(FontLoaderService, lambda: FontLoaderService())
     binder.bind_to_constructor(PlayerService, lambda: PlayerService())
