+++ mpvqc/pyobjects/document_exporter.py
@@ -16,59 +16,21 @@
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-from pathlib import Path
-from typing import Optional
-
 import inject
 from PySide6.QtCore import QObject, Slot
-from PySide6.QtGui import QGuiApplication, QStandardItemModel
 from PySide6.QtQml import QmlElement
-from PySide6.QtWidgets import QApplication
 
-from mpvqc.services import SettingsService, PlayerService
-from .comment_model import MpvqcCommentModelPyObject
+from mpvqc.services import DocumentExporterService
 
 QML_IMPORT_NAME = "pyobjects"
 QML_IMPORT_MAJOR_VERSION = 1
 
 
-def _find_comment_model_in_object_tree() -> MpvqcCommentModelPyObject:
-    engine = QGuiApplication.instance().engine
-    assert engine, "Cannot find QQmlApplicationEngine in QGuiApplication.instance()"
-    root = engine.rootObjects()
-    assert root, f"Cannot find root object in QQmlApplicationEngine"
-    model = root[0].findChild(QStandardItemModel, "mpvqcCommentModel")
-    assert model, f"Cannot find comment model in root object"
-    return model
-
-
 @QmlElement
 class MpvqcDocumentExporterPyObject(QObject):
-    _settings: SettingsService = inject.attr(SettingsService)
-    _player: PlayerService = inject.attr(PlayerService)
+    _exporter = inject.attr(DocumentExporterService)
 
-    def __init__(self):
-        super().__init__()
-        self._comment_model: Optional[MpvqcCommentModelPyObject] = None
-
-    @property
-    def _comments(self):
-        if self._comment_model is None:
-            self._comment_model = _find_comment_model_in_object_tree()
-        return self._comment_model.comments()
-
-    @Slot(result=str)
+    @Slot()
     def create_file_content(self) -> str:
-        for comment in self._comments:
-            print("py: comment", comment)
-
-        print("py: writeHeaderDate:", self._settings.writeHeaderDate)
-        print("py: writeHeaderGenerator:", self._settings.writeHeaderGenerator)
-        print("py: writeHeaderVideoPath:", self._settings.writeHeaderVideoPath)
-        print("py: writeHeaderNickname:", self._settings.writeHeaderNickname)
-
-        print("py: nickname:", self._settings.nickname)
-        print("py: generator:", f"{QApplication.applicationName()} {QApplication.applicationVersion()}")
-        print("py: video:", Path(self._player.mpv.path) if self._player.mpv.path else "")
-
-        return 'return value'
+        self._exporter.write()
+        return 'todo: remove-me'
