+++ mpvqc/services/__init__.py
@@ -18,6 +18,7 @@
 from .application_environment import ApplicationEnvironmentService
 from .application_paths import ApplicationPathsService
 from .backup import BackupService
+from .document_exporter import DocumentExporterService
 from .file_startup import FileStartupService
 from .font_loader import FontLoaderService
 from .operating_system_zoom_detector import OperatingSystemZoomDetectorService
